Separate lz77 module into libflate_lz77 crate
