+++ libflate_lz77/src/default.rs
@@ -177,28 +177,3 @@ impl LargePrefixTable {
         None
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use deflate::symbol::Symbol;
-
-    #[test]
-    // See: https://github.com/sile/libflate/issues/21
-    fn issue21() {
-        let mut enc = DefaultLz77Encoder::new();
-        let mut sink = Vec::new();
-        enc.encode(b"aaaaa", &mut sink);
-        enc.flush(&mut sink);
-        assert_eq!(
-            sink,
-            vec![
-                Symbol::Literal(97),
-                Symbol::Share {
-                    length: 4,
-                    distance: 1
-                }
-            ]
-        );
-    }
-}
